undo: Move redo execution under data.undo
Command checkpoint handling now lives in the data.undo.checkpoints module
inside the package. Undo and redo execution still remain in the top-level
data.undo_checkpoints module.

The redo command can use the packaged snapshot, conflict, and restoration
functions directly. Keeping redo at the old module path leaves the same
public operation in two architectural locations.

This commit moves redo execution into data.undo.checkpoints, updates the
redo command, and leaves an import at the old path for callers that have
not moved yet.

The next commit will exercise redo through the package before undo
execution removes data.undo_checkpoints.
